Refactoring of DBSpecificOption:
Remove the second unused constructor and change the type of the 'value'
field to String as it cannot be any other Object by now. The other changes are mainly code style changes.
As #getValue returns only String then the callers do not need to call
toString on the returned values.